+++ src/main/java/de/is24/deadcode4j/analyzer/SuperClassAnalyzer.java
@@ -5,7 +5,6 @@ import de.is24.deadcode4j.AnalysisContext;
 import de.is24.deadcode4j.analyzer.javassist.ClassPathFilter;
 import de.is24.guava.NonNullFunction;
 import javassist.CtClass;
-import javassist.NotFoundException;
 
 import javax.annotation.Nonnull;
 import java.util.Collections;
@@ -14,6 +13,8 @@ import java.util.Set;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.collect.Lists.newArrayList;
+import static de.is24.deadcode4j.analyzer.javassist.CtClasses.getSuperclassOf;
+import static de.is24.deadcode4j.analyzer.javassist.CtClasses.isJavaLangObject;
 
 /**
  * Serves as a base class with which to mark classes as being in use if they are a direct subclass of one of the
@@ -62,28 +63,23 @@ public abstract class SuperClassAnalyzer extends ByteCodeAnalyzer {
         if (knownSuperClasses.isEmpty()) {
             return;
         }
+
         String clazzName = clazz.getName();
         analysisContext.addAnalyzedClass(clazzName);
-        final List<String> classHierarchy;
-        try {
-            classHierarchy = getClassHierarchy(clazz);
-        } catch (NotFoundException e) {
-            logger.warn("The class path is not correctly set up; could not load [{}]! Skipping superclass check for {}.", e.getMessage(), clazzName);
-            return;
-        }
-        if (!Collections.disjoint(knownSuperClasses, classHierarchy)) {
+
+        if (!Collections.disjoint(knownSuperClasses, getClassHierarchy(clazz))) {
             analysisContext.addDependencies(this.dependerId, clazzName);
         }
     }
 
     @Nonnull
-    private List<String> getClassHierarchy(@Nonnull final CtClass clazz) throws NotFoundException {
+    private List<String> getClassHierarchy(@Nonnull final CtClass clazz) {
         List<String> classes = newArrayList();
         CtClass loopClass = clazz;
         do {
             classes.add(loopClass.getClassFile2().getSuperclass());
-            loopClass = loopClass.getSuperclass();
-        } while (loopClass != null && !"java.lang.Object".equals(loopClass.getName()));
+            loopClass = getSuperclassOf(loopClass);
+        } while (loopClass != null && !isJavaLangObject(loopClass));
         return classes;
     }