+++ src/main/java/de/is24/javaparser/ImportDeclarations.java
@@ -6,6 +6,8 @@ import japa.parser.ast.ImportDeclaration;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 /**
  * Provides convenience methods for dealing with {@link japa.parser.ast.ImportDeclaration}s.
  *
@@ -25,8 +27,9 @@ public final class ImportDeclarations {
     public static Predicate<? super ImportDeclaration> isAsterisk() {
         return new Predicate<ImportDeclaration>() {
             @Override
+            @SuppressWarnings("ConstantConditions")
             public boolean apply(@Nullable ImportDeclaration input) {
-                return input != null && input.isAsterisk();
+                return checkNotNull(input).isAsterisk();
             }
         };
     }
@@ -41,8 +44,9 @@ public final class ImportDeclarations {
     public static Predicate<? super ImportDeclaration> isStatic() {
         return new Predicate<ImportDeclaration>() {
             @Override
+            @SuppressWarnings("ConstantConditions")
             public boolean apply(@Nullable ImportDeclaration input) {
-                return input != null && input.isStatic();
+                return checkNotNull(input).isStatic();
             }
         };
     }