+++ src/test/java/de/is24/deadcode4j/analyzer/A_TypeErasureAnalyzer.java
@@ -50,4 +50,29 @@ public final class A_TypeErasureAnalyzer extends AnAnalyzer<TypeErasureAnalyzer>
                 "java.util.Collection");
     }
 
+    @Test
+    public void recognizesInheritedType() {
+        analyzeFile("../../src/test/java/de/is24/deadcode4j/analyzer/typeerasure/ClassWithInheritedType.java");
+
+        assertThatDependenciesAreReportedFor("de.is24.deadcode4j.analyzer.typeerasure.ClassWithInheritedType",
+                "de.is24.deadcode4j.analyzer.typeerasure.TypedArrayList$InnerClass");
+    }
+
+    @Test
+    public void recognizesAnonymousClasses() {
+        analyzeFile("../../src/test/java/de/is24/deadcode4j/analyzer/typeerasure/ClassWithAnonymousClasses.java");
+
+        assertThatDependenciesAreReportedFor("de.is24.deadcode4j.analyzer.typeerasure.ClassWithAnonymousClasses$1",
+                "java.lang.String");
+        assertThatDependenciesAreReportedFor("de.is24.deadcode4j.analyzer.typeerasure.ClassWithAnonymousClasses$1$1",
+                "java.lang.Integer");
+        assertThatDependenciesAreReportedFor("de.is24.deadcode4j.analyzer.typeerasure.ClassWithAnonymousClasses$1$2",
+                "de.is24.deadcode4j.junit.SomeInterface");
+        assertThatDependenciesAreReportedFor(
+                "de.is24.deadcode4j.analyzer.typeerasure.ClassWithAnonymousClasses$1$1AnonymousInner",
+                "java.lang.String");
+        assertThatDependenciesAreReportedFor("de.is24.deadcode4j.analyzer.typeerasure.ClassWithAnonymousClasses$2",
+                "java.util.Set");
+    }
+
 }